refactor(carousel): extract Slide to reusable component
Simplified CarouselPosterMini by extracting the Slide functionality into a new reusable component, CarouselPosterSlide.

- **src/components/carousel/variants/CarouselPosterMini.tsx**:
  - Replaced `Slide` with the newly extracted `CarouselPosterSlide`.
  - Removed the in-file `Slide` implementation.
  - Cleaned up imports to properly import the new `CarouselPosterSlide`.

- **src/components/carousel/components/CarouselPosterMIniSlide.tsx**:
  - Created a new component to encapsulate the presentation logic for carousel slides, reusing the extracted code from the old `Slide` implementation.
