+++ src/components/carousel/variants/CarouselPosterMini.tsx
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
-// @mui
-import Box from '@mui/material/Box';
-import Carousel, { useCarousel } from '@src/components/carousel/index';
 // @ts-ignore
-import { type Post } from '@lens-protocol/api-bindings/dist/declarations/src/lens/graphql/generated';
-import PosterHorizontal from '@src/components/poster/variants/poster-horizontal.tsx';
+import {Post} from "@lens-protocol/api-bindings/dist/declarations/src/lens/graphql/generated";
+
+import Box from '@mui/material/Box';
+
+import Carousel, { useCarousel } from '@src/components/carousel/index';
 import CarouselNavigationArrows from '@src/components/carousel/components/CarouselNavigationArrows.tsx';
 import { CarouselSection } from '@src/components/poster/carousel-section.tsx';
-
-// Types
-import { CarouselPosterMiniProps, CarouselPosterSlideProps } from './types';
 import { useItemsPerSlide } from '@src/hooks/components/use-item-per-slide.ts';
+import CarouselPosterSlide from "@src/components/carousel/components/CarouselPosterMIniSlide.tsx";
+
+import { CarouselPosterMiniProps } from './types';
 
 // ----------------------------------------------------------------------
 
@@ -74,7 +74,7 @@ export default function CarouselPosterMini({
       >
         <Carousel ref={carousel.carouselRef} {...carousel.carouselSettings}>
           {slideData.map((slideItems, index) => (
-            <Slide
+            <CarouselPosterSlide
               key={`slide-publications-${index}`}
               items={slideItems}
               itemsPerRow={itemsPerSlide}
@@ -83,48 +83,5 @@ export default function CarouselPosterMini({
         </Carousel>
       </Box>
     </CarouselSection>
-  );
-}
-
-function Slide({ items, itemsPerRow }: CarouselPosterSlideProps) {
-  const row1 = items.slice(0, itemsPerRow);
-  const row2 = items.slice(itemsPerRow, itemsPerRow * 2);
-  const itemWidthPercent = 100 / itemsPerRow;
-
-  // const getMediaUri = (cid: string): string => `https://ipfs.io/ipfs/${cid?.replace?.('ipfs://', '')}`
-  const getMediaUri = (cid: string): string => `${cid}`;
-  const getWallpaperCid = (post: any): string =>
-    post?.metadata?.attachments?.find((el: any) => el.altTag === 'wallpaper')?.image?.raw?.uri;
-  const getPosterCid = (post: any): string =>
-    post?.metadata?.attachments?.find((el: any) => el.altTag === 'poster')?.image?.raw?.uri;
-
-  return (
-    <Box>
-      {[row1, row2].map((rowItems, rowIndex) => (
-        <Box key={`row-publications-${rowIndex}`} sx={{ display: 'flex' }}>
-          {rowItems.map((post) => (
-            <Box
-              key={post.id}
-              sx={{
-                flexBasis: `${itemWidthPercent}%`,
-                maxWidth: `${itemWidthPercent}%`,
-                p: 1,
-              }}
-            >
-              <PosterHorizontal
-                id={post.id}
-                title={post.metadata.title}
-                images={{
-                  vertical: getMediaUri(getPosterCid(post)),
-                  wallpaper: getMediaUri(getWallpaperCid(post)),
-                }}
-                likes={post.globalStats.upvotes}
-                synopsis={post.metadata.content}
-              />
-            </Box>
-          ))}
-        </Box>
-      ))}
-    </Box>
   );
 }
